reduce cycles to 2
