+++ src/core/resources/assets/index.ts
@@ -1,5 +1,3 @@
 export { useAssets } from './assets';
-export { useUserAssets } from './userAssets';
-export type { UserAssetsArgs } from './userAssets';
-export { useUserAssetsByChain } from './userAssetsByChain';
-export type { UserAssetsByChainArgs } from './userAssetsByChain';
+export { useUserAssets, useUserAssetsByChain } from './userAssets';
+export type { UserAssetsArgs, UserAssetsByChainArgs } from './userAssets';
